fix(records-api): align id_cliente to Integer and relax schema validation
- id_cliente is INT across the whole schema (legacy cliente PK + medida_* tables)
  but the kernel mapped it as Long. Aligned to Integer in c4e-persistence-core
  (Cliente, AbstractMeasure, MedidaCCH, BaseClienteRepository) plus the ripple in
  records-api and ingestion-service (repos, service, controller, adapter, tests).
- records-api ddl-auto: validate -> none. Measure value columns are INT UNSIGNED
  (range exceeds signed int) and are deliberately mapped as Long to avoid overflow;
  Hibernate validate rejects Long<->INT UNSIGNED though runtime read/write works.
  db-migrations owns the schema, so the app trusts the migrations instead.
- ClienteInfoDTO converted to a record (less boilerplate, immutable).

+++ c4e-ingestion-service/src/main/java/com/com4energy/processor/service/measure/persistence/JpaMeasurePersistenceAdapter.java
@@ -284,7 +284,7 @@ public class JpaMeasurePersistenceAdapter implements MeasurePersistenceContracts
 
     private MedidaH toMedidaH(
             MeasureRecord.Hourly measure,
-            Long clientId,
+            Integer clientId,
             MeasurePersistenceContracts.PersistMeasuresCommand command
     ) {
         return MedidaH.builder()
@@ -317,7 +317,7 @@ public class JpaMeasurePersistenceAdapter implements MeasurePersistenceContracts
 
     private MedidaQH toMedidaQH(
             MeasureRecord.QuarterHourly measure,
-            Long clientId,
+            Integer clientId,
             MeasurePersistenceContracts.PersistMeasuresCommand command
     ) {
         return MedidaQH.builder()
@@ -369,7 +369,7 @@ public class JpaMeasurePersistenceAdapter implements MeasurePersistenceContracts
         }
     }
 
-    private MedidaCCH toMedidaCch(MeasureRecord.Cch measure, Long clientId) {
+    private MedidaCCH toMedidaCch(MeasureRecord.Cch measure, Integer clientId) {
         return MedidaCCH.builder()
                 .clienteId(clientId)
                 .fecha(measure.timestamp())
@@ -379,9 +379,9 @@ public class JpaMeasurePersistenceAdapter implements MeasurePersistenceContracts
                 .build();
     }
 
-    private record ClienteResolution(Long clientId, String tarifa, String errorMessage) {
+    private record ClienteResolution(Integer clientId, String tarifa, String errorMessage) {
 
-        static ClienteResolution success(Long clientId, String tarifa) {
+        static ClienteResolution success(Integer clientId, String tarifa) {
             return new ClienteResolution(clientId, tarifa, null);
         }
